dep bumps

+++ generators/uimodule/ui5Libs.js
@@ -105,7 +105,7 @@ export default class extends Generator {
 		fs.writeFileSync(this.destinationPath("webapp/manifest.json"), JSON.stringify(manifestJSON, null, 4))
 
 		// remove option to bootstrap from local UI5 sources, as UI5 source is part of user selection
-		fs.unlinkSync(this.destinationPath("ui5-local.yaml"))
+		if (fs.existsSync(this.destinationPath("ui5-local.yaml"))) fs.unlinkSync(this.destinationPath("ui5-local.yaml"))
 		delete packageJson.scripts["start-local"]
 		fs.unlinkSync(this.destinationPath("package.json")) // avoid conflict/auto-prompt by yeoman
 		fs.writeFileSync(this.destinationPath("package.json"), JSON.stringify(packageJson, null, 4))
